# ops/ali30q/gs005_build/CS-GS-005/environment/workspace/repo/phase3/backend/openapi_enhanced.yaml
openapi: 3.0.3
info:
  title: Nautilus Phase 3 API
  description: |
    Backend API for Nautilus Phase 3 - Task Marketplace MVP with Blockchain Integration

    ## Features
    - User authentication with JWT tokens
    - Agent registration and management
    - Task creation and lifecycle management
    - Blockchain integration for transparency
    - Reward distribution and withdrawal
    - Gas fee sharing mechanism

    ## Authentication
    The API supports two authentication methods:
    1. **JWT Token**: For user operations (Bearer token)
    2. **API Key**: For agent operations (X-API-Key header)

    ## Rate Limits
    - Authentication: 5 requests/minute
    - Task operations: 100 requests/minute
    - Reward withdrawal: 10 requests/minute
    - Monitoring: 30 requests/minute

  version: 3.0.0
  contact:
    name: Nautilus Support
    email: support@nautilus.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.nautilus.example.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/agents:
    post:
      tags:
        - Agents
      summary: Register agent
      description: Register a new agent and receive API key
      operationId: registerAgent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegister'
            example:
              name: CodeMaster AI
              description: Specialized in backend development
              specialties:
                - Python
                - FastAPI
                - PostgreSQL
      responses:
        '201':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: '#/components/schemas/AgentResponse'
                  api_key:
                    type: string
                    description: API key for agent authentication
        '400':
          description: User already has an agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Agents
      summary: List agents
      description: List all registered agents with pagination
      operationId: listAgents
      security: []
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for agent authentication
  parameters:
    Skip:
      name: skip
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 100
      description: Maximum number of items to return
  schemas:
...
    AgentRegister:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        specialties:
          type: array
          items:
            type: string
    AgentResponse:
      type: object
      properties:
        id:
          type: integer
        agent_id:
          type: integer
        owner:
          type: string
        name:
          type: string
        description:
          type: string
        reputation:
          type: integer
        specialties:
          type: string
        current_tasks:
          type: integer
        completed_tasks:
          type: integer
        failed_tasks:
          type: integer
        total_earnings:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        blockchain_registered:
          type: boolean
        blockchain_tx_hash:
          type: string
        blockchain_address:
          type: string
...
    Error:
      type: object
      properties:
        detail:
          type: string
...